now creating an api key if one doesn't exist

+++ src/server/controllers/api-key.mjs
@@ -1,14 +1,10 @@
 import express from 'express';
 import mongoose from 'mongoose';
-import {
-    APIGatewayClient,
-    CreateApiKeyCommand, CreateUsagePlanKeyCommand,
-    DeleteApiKeyCommand
-} from '@aws-sdk/client-api-gateway';
+import {APIGatewayClient, DeleteApiKeyCommand} from '@aws-sdk/client-api-gateway';
 
 import {isAuthenticated} from '../middleware/authentication.mjs';
 
-const {AWS_API_GATEWAY_PLAN_ID} = process.env;
+const {ENVIRONMENT} = process.env;
 
 const client = new APIGatewayClient({region: 'us-east-2'});
 
@@ -31,46 +27,11 @@ ApiKeyRouter.get('/', async (req, res, next) => {
 });
 
 ApiKeyRouter.post('/', async (req, res, next) => {
-
     try {
         const ApiKey = mongoose.model('ApiKey');
-        const dioptraUserId = req.user._id;
-        const requestOrganization = req.user.requestOrganization;
-        const dioptraOrganizationId = requestOrganization._id;
-        const dioptraApiKey = new ApiKey({
-            user: req.user._id,
-            organization: requestOrganization._id
-        });
+        const dioptraApiKey = await ApiKey.createApiKeyForUser(req.user);
 
-        let awsApiKey = {
-            id: `__api_key_id__${Date.now()}__`,
-            value: `__api_key_value__${Date.now()}__`
-        };
-
-        try {
-            awsApiKey = await client.send(new CreateApiKeyCommand({
-                enabled: true,
-                name: `${dioptraApiKey._id} (as: ${req.user.username} | ${requestOrganization.name})`,
-                tags: {
-                    dioptraUserId, dioptraOrganizationId,
-                    dioptraApiKeyId: dioptraApiKey._id
-                }
-            }));
-
-            await client.send(new CreateUsagePlanKeyCommand({
-                keyId: awsApiKey.id,
-                keyType: 'API_KEY',
-                // TODO: Change this depending on what plan the organization is on when people are paying us.
-                usagePlanId: AWS_API_GATEWAY_PLAN_ID
-            }));
-        } catch (e) {
-            console.warn(`WARNING: Failed to create API key. This might be because we're running in development mode: ${e}`);
-        }
-
-        dioptraApiKey.awsApiKeyId = awsApiKey.id;
-        dioptraApiKey.awsApiKey = awsApiKey.value;
-
-        res.json(await dioptraApiKey.save());
+        res.json(dioptraApiKey);
     } catch (e) {
         next(e);
     }
@@ -85,12 +46,10 @@ ApiKeyRouter.delete('/:_id', async (req, res, next) => {
             user: req.user._id
         });
 
-        try {
+        if (ENVIRONMENT !== 'local-dev') {
             await client.send(new DeleteApiKeyCommand({
                 apiKey: dioptraApiKey.awsApiKeyId
             }));
-        } catch (e) {
-            console.warn(`WARNING: Failed to delete API key ${dioptraApiKey.awsApiKeyId}. This might be because we're running in development mode.`);
         }
 
         res.json(dioptraApiKey);
